requirements.txt
# Use Python 3.13

pytest==8.3.4
pytest-xdist==3.6.1
pytest-cov==6.0.0
snakeviz==2.2.2

pydantic==2.9.2
attrs==25.1.0
cattrs==24.1.2
pyyaml==6.0.2

tqdm==4.67.1
pqdm==0.2.0

hydra-core==1.3.2
fire==0.7.0
notebook==7.3.2

requests==2.32.3
bidict==0.23.1
maturin==1.8.1

numpy==2.2.2
pandas==2.2.3
pyarrow==19.0.0
tabulate==0.9.0
polars==1.21.0
scipy==1.15.1
scikit-learn==1.6.1
matplotlib==3.10.0
seaborn==0.13.2
plotnine==0.14.5

evaluate==0.4.3
datasets==3.2.0

openai==1.60.2
tiktoken==0.8.0

sql-metadata==2.6.0
sqlglot==26.0.1
duckdb==1.1.3

fastapi==0.115.7

faker==35.0.0
playwright==1.49.1
pytest-playwright==0.6.2
chromadb==0.3.25

ai-api-client-sdk==2.4.6
generative-ai-hub-sdk[amazon]==4.3.1
hana-ml==2.24.25040300